Extract search persistence and replay state
