+++ crates/truthdb-core/src/engine/mod.rs
@@ -21,33 +21,23 @@ pub(crate) use relational::{
 use std::collections::BTreeMap;
 use std::sync::{Arc, RwLock};
 
-use serde::{Deserialize, Serialize};
 use serde_json::{Value, json};
 use thiserror::Error;
 
 use crate::storage::{Storage, StorageError};
 pub use search::CommandError;
 use search::{
-    Command, Document, EngineState, FieldType, IndexState, SearchQuery, parse_command, render_json,
+    Command, ENGINE_WAL_ENTRY_TYPE, ENGINE_WAL_ENTRY_VERSION, EngineMeta, EngineState, SearchQuery,
+    WalEvent, decode_snapshot, parse_command, render_json,
 };
+#[cfg(test)]
+use search::{Document, FieldType, IndexState};
 
-const ENGINE_WAL_ENTRY_VERSION: u16 = 1;
-const ENGINE_WAL_ENTRY_TYPE: u16 = 1;
 const WAL_CHECKPOINT_THRESHOLD: f64 = 0.75;
 /// A standby takes a restartpoint earlier than a primary checkpoints (0.5 vs
 /// 0.75): its reclaim is capped at the shipped undo floor, so starting earlier
 /// leaves headroom when the floor lags.
 const STANDBY_RESTARTPOINT_THRESHOLD: f64 = 0.5;
-
-/// The engine's search-subsystem state, mutated only on the native path
-/// ([`Engine::execute`]) and read by the checkpointer. Guarded by
-/// [`Engine::meta`], a `RwLock` that doubles as the execution gate that keeps
-/// the two paths from observing each other's torn state (see [`Engine`]).
-struct EngineMeta {
-    state: EngineState,
-    next_seq_no: u64,
-    next_doc_id: u64,
-}
 
 /// The database engine, shared across the worker pool as `Arc<Engine>`.
 ///
@@ -538,79 +528,6 @@ impl Engine {
     }
 }
 
-impl EngineMeta {
-    fn apply_event(&mut self, event: &WalEvent) -> Result<(), EngineError> {
-        match event {
-            WalEvent::CreateIndex { name, mappings } => {
-                if self.state.indices.contains_key(name) {
-                    return Err(EngineError::Replay(format!(
-                        "wal attempted to recreate existing index '{name}'"
-                    )));
-                }
-                let index = IndexState::new(mappings.clone());
-                self.state.indices.insert(name.clone(), index);
-            }
-            WalEvent::InsertDocument {
-                index,
-                id,
-                document,
-            } => {
-                let index_state = self.state.indices.get_mut(index).ok_or_else(|| {
-                    EngineError::Replay(format!("wal references unknown index '{index}'"))
-                })?;
-                index_state.insert_document(id, document)?;
-                self.next_doc_id = self
-                    .next_doc_id
-                    .max(id.parse::<u64>().unwrap_or(0).saturating_add(1));
-            }
-        }
-        Ok(())
-    }
-
-    fn validate_create_index(
-        &self,
-        name: &str,
-        mappings: &BTreeMap<String, FieldType>,
-    ) -> Result<(), EngineError> {
-        if self.state.indices.contains_key(name) {
-            return Err(CommandError::IndexAlreadyExists(name.to_string()).into());
-        }
-        if mappings.is_empty() {
-            return Err(CommandError::InvalidCommand(
-                "index must define at least one mapped field".to_string(),
-            )
-            .into());
-        }
-        Ok(())
-    }
-
-    fn validate_insert_document(
-        &self,
-        index: &str,
-        document: &Document,
-    ) -> Result<(), EngineError> {
-        let index_state = self
-            .state
-            .indices
-            .get(index)
-            .ok_or_else(|| CommandError::UnknownIndex(index.to_string()))?;
-        index_state.validate_document(document)
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-enum WalEvent {
-    CreateIndex {
-        name: String,
-        mappings: BTreeMap<String, FieldType>,
-    },
-    InsertDocument {
-        index: String,
-        id: String,
-        document: Document,
-    },
-}
-
 #[derive(Debug, Error)]
 pub enum EngineError {
     #[error("{0}")]
@@ -673,21 +590,6 @@ fn render_sql_outcome(outcome: &crate::engine::BatchOutcome) -> String {
         })
     });
     json!({ "kind": "sql", "results": rendered, "error": error }).to_string()
-}
-
-/// Decodes a snapshot payload: JSON (current format), falling back to
-/// bincode for snapshots written by older versions. Legacy bincode snapshots
-/// can only have been document-free (bincode cannot deserialize
-/// `serde_json::Value`, so document-bearing ones were never loadable).
-fn decode_snapshot(data: &[u8]) -> Result<EngineState, EngineError> {
-    match serde_json::from_slice(data) {
-        Ok(state) => Ok(state),
-        Err(json_err) => bincode::deserialize(data).map_err(|bincode_err| {
-            EngineError::Replay(format!(
-                "failed to decode snapshot: as json: {json_err}; as legacy bincode: {bincode_err}"
-            ))
-        }),
-    }
 }
 
 #[cfg(test)]
