+++ crates/truthdb-core/src/engine/search/mod.rs
@@ -2,7 +2,13 @@
 
 mod command;
 mod index;
+mod persistence;
 
 pub use command::CommandError;
 pub(super) use command::{Command, parse_command, render_json};
-pub(super) use index::{Document, EngineState, FieldType, IndexState, SearchQuery};
+#[cfg(test)]
+pub(super) use index::{Document, FieldType, IndexState};
+pub(super) use index::{EngineState, SearchQuery};
+pub(super) use persistence::{
+    ENGINE_WAL_ENTRY_TYPE, ENGINE_WAL_ENTRY_VERSION, EngineMeta, WalEvent, decode_snapshot,
+};
