+++ crates/truthdb-core/src/engine/search/persistence.rs
@@ -1,0 +1,105 @@
+use std::collections::BTreeMap;
+
+use serde::{Deserialize, Serialize};
+
+use super::super::{CommandError, EngineError};
+use super::index::{Document, EngineState, FieldType, IndexState};
+
+pub(in crate::engine) const ENGINE_WAL_ENTRY_VERSION: u16 = 1;
+pub(in crate::engine) const ENGINE_WAL_ENTRY_TYPE: u16 = 1;
+/// The engine's search-subsystem state, mutated only on the native path
+/// ([`Engine::execute`]) and read by the checkpointer. Guarded by
+/// [`Engine::meta`], a `RwLock` that doubles as the execution gate that keeps
+/// the two paths from observing each other's torn state (see [`Engine`]).
+pub(in crate::engine) struct EngineMeta {
+    pub(in crate::engine) state: EngineState,
+    pub(in crate::engine) next_seq_no: u64,
+    pub(in crate::engine) next_doc_id: u64,
+}
+
+impl EngineMeta {
+    pub(in crate::engine) fn apply_event(&mut self, event: &WalEvent) -> Result<(), EngineError> {
+        match event {
+            WalEvent::CreateIndex { name, mappings } => {
+                if self.state.indices.contains_key(name) {
+                    return Err(EngineError::Replay(format!(
+                        "wal attempted to recreate existing index '{name}'"
+                    )));
+                }
+                let index = IndexState::new(mappings.clone());
+                self.state.indices.insert(name.clone(), index);
+            }
+            WalEvent::InsertDocument {
+                index,
+                id,
+                document,
+            } => {
+                let index_state = self.state.indices.get_mut(index).ok_or_else(|| {
+                    EngineError::Replay(format!("wal references unknown index '{index}'"))
+                })?;
+                index_state.insert_document(id, document)?;
+                self.next_doc_id = self
+                    .next_doc_id
+                    .max(id.parse::<u64>().unwrap_or(0).saturating_add(1));
+            }
+        }
+        Ok(())
+    }
+
+    pub(in crate::engine) fn validate_create_index(
+        &self,
+        name: &str,
+        mappings: &BTreeMap<String, FieldType>,
+    ) -> Result<(), EngineError> {
+        if self.state.indices.contains_key(name) {
+            return Err(CommandError::IndexAlreadyExists(name.to_string()).into());
+        }
+        if mappings.is_empty() {
+            return Err(CommandError::InvalidCommand(
+                "index must define at least one mapped field".to_string(),
+            )
+            .into());
+        }
+        Ok(())
+    }
+
+    pub(in crate::engine) fn validate_insert_document(
+        &self,
+        index: &str,
+        document: &Document,
+    ) -> Result<(), EngineError> {
+        let index_state = self
+            .state
+            .indices
+            .get(index)
+            .ok_or_else(|| CommandError::UnknownIndex(index.to_string()))?;
+        index_state.validate_document(document)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub(in crate::engine) enum WalEvent {
+    CreateIndex {
+        name: String,
+        mappings: BTreeMap<String, FieldType>,
+    },
+    InsertDocument {
+        index: String,
+        id: String,
+        document: Document,
+    },
+}
+/// Decodes a snapshot payload: JSON (current format), falling back to
+/// bincode for snapshots written by older versions. Legacy bincode snapshots
+/// can only have been document-free (bincode cannot deserialize
+/// `serde_json::Value`, so document-bearing ones were never loadable).
+pub(in crate::engine) fn decode_snapshot(data: &[u8]) -> Result<EngineState, EngineError> {
+    match serde_json::from_slice(data) {
+        Ok(state) => Ok(state),
+        Err(json_err) => bincode::deserialize(data).map_err(|bincode_err| {
+            EngineError::Replay(format!(
+                "failed to decode snapshot: as json: {json_err}; as legacy bincode: {bincode_err}"
+            ))
+        }),
+    }
+}
